feat: add continue shopping btn functionality to basket
The function handleContinueShoppingBtnClick toggles the visibility of the 'basket--content' and 'basket--empty' elements. If 'basket--content' is currently shown, it is hidden and marked as 'hidden'. If 'basket--empty' is currently shown, it is hidden and marked as 'hidden'.

This enhancement improves the user experience by providing an intuitive way to navigate back to the shopping page from the basket.