refactor: optimize handleGoToCartBtnClick function
This commit refactors the handleGoToCartBtnClick function to improve code readability and simplify the logic.

Instead of using classList.remove and classList.add to show and hide elements, this commit introduces the use of toggleAttribute method for toggling the visibility of the basket--content and basket--empty elements based on the length of the basketItemsArray.

By using toggleAttribute('data-visible'), the code becomes more concise and easier to understand. The basketContentEl and basketEmptyEl variables are introduced to improve code readability and avoid repetitive calls to document.getElementById.

These changes result in a more efficient and maintainable implementation of the handleGoToCartBtnClick function.